+++ stroom-cache/stroom-cache-api/src/main/java/stroom/cache/api/CacheManager.java
@@ -24,9 +24,11 @@ import java.util.function.Supplier;
 
 public interface CacheManager extends AutoCloseable {
 
-    <K, V> ICache<K, V> create(String name, Supplier<CacheConfig> cacheConfigSupplier);
+    <K, V> ICache<K, V> create(String name,
+                               Supplier<CacheConfig> cacheConfigSupplier);
 
-    <K, V> ICache<K, V> create(String name, Supplier<CacheConfig> cacheConfigSupplier, Function<K, V> loadFunction);
+    <K, V> ICache<K, V> create(String name,
+                               Supplier<CacheConfig> cacheConfigSupplier, Function<K, V> loadFunction);
 
     <K, V> ICache<K, V> create(String name,
                                Supplier<CacheConfig> cacheConfigSupplier,