+++ stroom-data/stroom-data-store-impl-fs/src/main/java/stroom/data/store/impl/fs/FsVolumeListUtil.java
@@ -26,7 +26,7 @@ public class FsVolumeListUtil {
     public static List<FsVolume> removeFullVolumes(final List<FsVolume> list) {
         final List<FsVolume> newList = new ArrayList<>(list.size());
         for (final FsVolume volume : list) {
-            if (!volume.isFull()) {
+            if (!volume.getCapacityInfo().isFull()) {
                 newList.add(volume);
             }
         }