+++ stroom-util/src/main/java/stroom/util/xml/MyXmlEmitter.java
@@ -190,6 +190,7 @@ public class MyXmlEmitter extends XMLEmitter {
      * @param c The char to test.
      * @return True if the char is forbidden.
      */
+    @SuppressWarnings("checkstyle:AvoidEscapedUnicodeCharacters")
     private boolean isForbidden(final char c) {
         // surrogates, U+FFFE and U+FFFF are forbidden in XML.
         return c == INVALID_CHARACTER_1 || c == INVALID_CHARACTER_2;