+++ stroom-util/src/test/java/stroom/util/io/capacity/TestRandomCapacitySelector.java
@@ -1,0 +1,11 @@
+package stroom.util.io.capacity;
+
+import org.junit.jupiter.api.Test;
+
+class TestRandomCapacitySelector extends AbstractHasCapacitySelectorTest {
+
+    @Test
+    void test() {
+        testMultipleTimes(new RandomCapacitySelector(), ALL_PATHS);
+    }
+}